[DMA] re-organize the ShotDetect.cs and fix some bugs
