+++ ShotsDetect/ShotsDetect.csproj
@@ -48,6 +48,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DetectMethod.cs" />
     <Compile Include="ShotsXml.cs" />
     <Compile Include="ShotsDetect.cs" />
     <Compile Include="Form1.cs">
